requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-openai==0.0.8
python-dotenv==1.0.1
chromadb==0.6.3
pypdf==4.1.0
tiktoken==0.6.0
watchdog==3.0.0 